requirements.txt
langchain-community>=0.0.10
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
pathlib>=1.0.1
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
drf-yasg>=1.21.7
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9 